Introduce a "traces" package; move dstat runner there.
